--- a/BarcodeStandard/BarcodeCommon.cs
+++ b/BarcodeStandard/BarcodeCommon.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace BarcodeStandard
 {
@@ -17,15 +16,14 @@ namespace BarcodeStandard
 
         internal static bool CheckNumericOnly(string data)
         {
-            char c;
-            for (int i = 0; i < data.Length; i++)
+            foreach (var c in data)
             {
-                c = data[i];
                 if (c < '0' && c > '9') return false;
             }
+
             return true;
         }
-        
+
         internal static int GetAlignmentShiftAdjustment(Barcode barcode)
         {
             switch (barcode.Alignment)